refactor(node): separate configuration and observability owners
Split configuration into private domain modules with explicit public exports.
Consolidate option overlays, keep final-network mining address resolution,
and move journal validation beside its settings. Separate metrics serving,
warning state, uptime, and ledger evidence without changing public paths.
Preserve existing tests and add ten boundary regressions.
