+++ crates/node/src/config/auth.rs
@@ -1,0 +1,76 @@
+//! RPC credentials and their redacted diagnostic representation.
+
+use core::fmt;
+use std::path::PathBuf;
+
+use anyhow::Result;
+
+const DEFAULT_RPC_USER: &str = "bitcoin-rs";
+const DEFAULT_RPC_PASSWORD: &str = "bitcoin-rs";
+
+/// RPC authentication configuration.
+#[derive(Clone, Eq, PartialEq)]
+pub enum Auth {
+    /// HTTP Basic credentials.
+    Basic {
+        /// RPC username.
+        user: String,
+        /// RPC password.
+        password: String,
+    },
+    /// Bitcoin Core cookie-auth file.
+    Cookie {
+        /// Cookie file path.
+        path: PathBuf,
+    },
+}
+
+impl Auth {
+    /// Constructs Basic authentication credentials.
+    #[must_use]
+    pub fn basic(user: impl Into<String>, password: impl Into<String>) -> Self {
+        Self::Basic {
+            user: user.into(),
+            password: password.into(),
+        }
+    }
+
+    /// Converts this configuration into the RPC crate's runtime auth policy.
+    pub fn to_rpc_auth(&self) -> Result<bitcoin_rs_rpc::Auth> {
+        match self {
+            Self::Basic { user, password } => {
+                Ok(bitcoin_rs_rpc::Auth::basic(user.clone(), password))
+            }
+            Self::Cookie { path } => Ok(bitcoin_rs_rpc::Auth::cookie(path)?),
+        }
+    }
+
+    pub(super) fn basic_parts(&self) -> (String, String) {
+        match self {
+            Self::Basic { user, password } => (user.clone(), password.clone()),
+            Self::Cookie { .. } => (DEFAULT_RPC_USER.to_owned(), DEFAULT_RPC_PASSWORD.to_owned()),
+        }
+    }
+}
+
+impl fmt::Debug for Auth {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::Basic { user, .. } => f
+                .debug_struct("Auth::Basic")
+                .field("user", user)
+                .field("password", &"<redacted>")
+                .finish(),
+            Self::Cookie { .. } => f
+                .debug_struct("Auth::Cookie")
+                .field("path", &"<redacted>")
+                .finish(),
+        }
+    }
+}
+
+impl Default for Auth {
+    fn default() -> Self {
+        Self::basic(DEFAULT_RPC_USER, DEFAULT_RPC_PASSWORD)
+    }
+}
