+++ crates/node/src/config/index.rs
@@ -1,0 +1,52 @@
+//! Script-index capability selection and legacy configuration spellings.
+
+/// How much of the derived `ScriptIndex` a node maintains.
+///
+/// `ScriptIndex` is rebuildable derived state, so the mode is a capability
+/// selection rather than a storage compatibility question: `full` adds
+/// historical funding/spending rows while still maintaining the live-output
+/// view.
+///
+/// The boolean spellings remain behaviorally compatible: `--scriptindex`,
+/// `--scriptindex=true`, and `BITCOIN_RS_SCRIPTINDEX=true` all mean
+/// [`Self::Full`], and `false` means [`Self::Disabled`].
+///
+/// [`Self::Utxo`] is the live-only selection; see `IDX-01`.
+#[derive(Copy, Clone, Debug, Default, Eq, PartialEq)]
+pub enum ScriptIndexMode {
+    /// No `ScriptIndex` capability is maintained.
+    #[default]
+    Disabled,
+    /// Compact live-output view only (`IDX-01`).
+    Utxo,
+    /// Maintain both the live-output view and historical script activity.
+    Full,
+}
+
+impl ScriptIndexMode {
+    /// Whether any `ScriptIndex` capability is enabled.
+    #[must_use]
+    pub const fn is_enabled(self) -> bool {
+        !matches!(self, Self::Disabled)
+    }
+
+    /// Whether historical funding/spending rows are maintained.
+    #[must_use]
+    pub const fn keeps_history(self) -> bool {
+        matches!(self, Self::Full)
+    }
+
+    /// Parses a mode from a configuration value.
+    ///
+    /// Accepts the historical boolean spellings for compatibility: `true`
+    /// means `full` and `false` means disabled. Parsing is case-insensitive.
+    #[must_use]
+    pub fn parse(value: &str) -> Option<Self> {
+        match value.trim().to_ascii_lowercase().as_str() {
+            "utxo" => Some(Self::Utxo),
+            "full" | "true" | "1" | "yes" => Some(Self::Full),
+            "false" | "0" | "no" => Some(Self::Disabled),
+            _ => None,
+        }
+    }
+}
